Develop

+++ Syndra/src/Engine/ImGui/ImGuiLayer.cpp
@@ -2,7 +2,7 @@
 #include "ImGuiLayer.h"
 #include "GLFW/glfw3.h"
 #include "glad/glad.h"
-#include "Engine/Application.h"
+#include "Engine/Core/Application.h"
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_opengl3.h"
 
@@ -76,7 +76,7 @@ namespace  Syndra {
 
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
